progress

+++ news/lib/src/resources/news_api_provider.dart
@@ -6,15 +6,14 @@ import 'repository.dart';
 
 final _root = 'https://hacker-news.firebaseio.com/v0';
 
-class NewsApiProvider implements Source{
+class NewsApiProvider implements Source {
   Client client = Client();
-
 
   Future<List<int>> fetchTopIds() async {
     final response = await client.get('$_root/topstories.json');
     final ids = json.decode(response.body);
 
-    return ids;
+    return ids.cast<int>();
   }
 
   Future<ItemModel> fetchItem(int id) async {
